fix: PR improvements
